+++ packages/core/CHANGELOG.md
@@ -2,3 +2,23 @@
 
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
+
+# [5.0.0-alpha.4](https://git.sami.int.thomsonreuters.com/elf/refinitiv-ui/compare/@refinitiv-ui/core@5.0.0-alpha.3...@refinitiv-ui/core@5.0.0-alpha.4) (2021-06-09)
+
+**Note:** Version bump only for package @refinitiv-ui/core
+
+
+
+
+
+# 5.0.0-alpha.3 (2021-06-09)
+
+**Note:** Version bump only for package @refinitiv-ui/core
+
+
+
+
+
+# 5.0.0-alpha.2 (2021-06-09)
+
+**Note:** Version bump only for package @refinitiv-ui/core